feat: add vercel serverless setup
